refactor(models): 拆分 academic eams 模型
将 academic_eams.dart 拆分为 barrel 入口和多个领域子模型文件，保持现有导入路径不变。\n\nCo-authored-by: Codex <noreply@openai.com>
